Make logging behave
This PR removes all custom logging magic, and changes the behaviour to
the standard way of logging for Python libraries.

In short, instead of magic global loggers and setting global
configuration (so that applications experience problems), all logging
calls are simply emitted on module-level loggers. It's up to the
application to route these messages somewhere useful by configuring
handlers and setting levels.

Fixes #11.