refactor: rename the first argument passed to .api.multi_load
rename the first argument passed to .api.single_load from 'paths' to
'inputs' because now it might not be only a list of path strings but
also be a list of file / file-like objects, pathlib.Path objects or a
namedtuple `.inputs.Input` objects to load data from.

along with that, update its doctext to apply that change.