enhancement: add .utils.is_io_obj and .utils.is_output_obj
add .utils.is_io_obj to check if given object is a namedtuple 'Input' or
'Output', and .utils.is_output_obj to check if given object is a
namedtuple 'Output'. Along with that all calls of .utils.is_input_obj
are basically replaced with calls of .utils.is_io_obj.