change: [doc] remove the changelog in anyconfig.backend.base
I decided to remove all of the changelog entries in each modules
basically to simplify the code. I'll note in the top level changelog
like sections as muhc as possible if such changes in each modules to
nofity or warn users.

This change is for anyconfig.backend.base as a starting point.